requirements.txt
fastapi==0.75.0
pydantic==1.10.2
uvicorn==0.15.0
pymongo==3.12.1
motor==2.5.0
python-dotenv==0.19.2
